% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Used Cars data-dictionary}
\format{
A \code{tibble} containing 3 columns and 17 rows (one for each column of \code{usedcars} dataset)
\describe{
\item{columns}{Name of the column}
\item{short_definition}{Information of the contents of the column}
\item{column_order}{Column number in the dataset}
}
}
\usage{
dictionary
}
\description{
This dataset gives additional information about the columns in \code{usedcars} dataset.
}
\keyword{datasets}
