% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_model.R
\docType{data}
\name{rf_model}
\alias{rf_model}
\title{Random Forest Model for Used Cars Price Prediction}
\format{
An object of class \code{randomForest.formula} (inherits from \code{randomForest}) of length 18.
}
\usage{
rf_model
}
\description{
This is a Random Forest model created using bootstrapped data sampling with ntree=300 and mtry=4
and formula = listingPrice ~ variant + year + odometerReading + fuelType + bodyType + transmission
You can see the detailed code here (https://github.com/vkatti/ResearchProjectMBA/tree/main/07_ML_Modelling)
}
\keyword{datasets}
